cleanup LoRa Interface
