+++ lora_constants.py
@@ -1,0 +1,54 @@
+#Constants
+FLAGS_ACK = 0x80
+BROADCAST_ADDRESS = 255
+# 
+REG_00_FIFO = 0x00
+REG_01_OP_MODE = 0x01
+REG_06_FRF_MSB = 0x06
+REG_07_FRF_MID = 0x07
+REG_08_FRF_LSB = 0x08
+REG_0E_FIFO_TX_BASE_ADDR = 0x0e
+REG_0F_FIFO_RX_BASE_ADDR = 0x0f
+REG_10_FIFO_RX_CURRENT_ADDR = 0x10
+REG_12_IRQ_FLAGS = 0x12
+REG_13_RX_NB_BYTES = 0x13
+REG_1D_MODEM_CONFIG1 = 0x1d
+REG_1E_MODEM_CONFIG2 = 0x1e
+REG_19_PKT_SNR_VALUE = 0x19
+REG_1A_PKT_RSSI_VALUE = 0x1a
+REG_20_PREAMBLE_MSB = 0x20
+REG_21_PREAMBLE_LSB = 0x21
+REG_22_PAYLOAD_LENGTH = 0x22
+REG_26_MODEM_CONFIG3 = 0x26
+
+REG_4D_PA_DAC = 0x4d
+REG_40_DIO_MAPPING1 = 0x40
+REG_0D_FIFO_ADDR_PTR = 0x0d
+
+PA_DAC_ENABLE = 0x07
+PA_DAC_DISABLE = 0x04
+PA_SELECT = 0x80
+
+CAD_DETECTED_MASK = 0x01
+RX_DONE = 0x40
+TX_DONE = 0x08
+CAD_DONE = 0x04
+CAD_DETECTED = 0x01
+
+LONG_RANGE_MODE = 0x80
+MODE_SLEEP = 0x00
+MODE_STDBY = 0x01
+MODE_TX = 0x03
+MODE_RXCONTINUOUS = 0x05
+MODE_CAD = 0x07
+
+REG_09_PA_CONFIG = 0x09
+FXOSC = 32000000.0
+FSTEP = (FXOSC / 524288)
+
+# Hardware Constants
+LORA_SPI_CHANNEL = const(1)
+LORA_SCK  = const(10)
+LORA_MOSI = const(11)
+LORA_MISO = const(8)
+LORA_CS   = const(9)
